Make mentor cards generic to be used by both type of users
