+++ lib/l10n/app_en.arb
@@ -30,16 +30,10 @@
     "homeInvitationSectionTitle": "Your new invitations",
     "homeInvitationRequested": "New Request",
     "homeInvitationAccepted": "Request Accepted",
+    "homeRecommendedEntrepreneurTitle": "Mentors for you",
+    "homeRecommendedMentorTitle": "Entrepreneurs for you",
+    "homeRecommendedSubtitle": "Recommendations are based on profile",
     "percentageProfileCompletion": "{profileCompletionPercentage}%",
-    "homeMentorsSectionTitle": "Recommended Mentors",
-    "homeResourcesSectionTitle": "Recommended Tools",
-    "homeResourcesWebinar": "Webinar",
-    "homeResourcesTraining": "Training",
-    "homeResourcesBlogs": "Blogs",
-    "homeResourcesWebinarDescription": "Mentoring First Step",
-    "homeResourcesTrainingDescription": "Masterclass Training",
-    "homeResourcesBlogsDescription": "More Information",
-    "homeFindMoreMentors": "Find More Mentors",
     "inboxInvitesReceived": "Received invites",
     "inboxInvitesReceivedMessage": "You have a new invitation to connect",
     "inboxInvitesSent": "Sent invites",
@@ -360,15 +354,10 @@
     "signUpVisibleInfoDesc" : "This will be shown on your profile.",
     "continueButton": "Continue",
     "findMentors": "Find Mentors",
-    "signUpCompletedTextOne": "You can view and edit your profile any time and even fill out additional fields.",
-    "signUpCompletedTextTwo": "Next, start connecting with mentors. We recommend you reach out to at least 3 mentors for your best chances of finding a good match.",
-    "signUpCompletedTextTwoMentor": "Next, start connecting with entrepreneurs. We recommend you reach out to at least 3 entrepreneurs for your best chances of finding a good match.",
     "moreResources": "More resources",
     "trainingMentoringFirstSteps": "Training: Mentoring First Steps",
     "trainingBuildingFuture": "Training: Building a Sustainable Future",
     "trainingMarketingBusiness": "Marketing Your Business 101",
-    "mentorsForYou": "Mentors for you",
-    "recommendationsBasedOnYourProfile": "Recommendations are based on profile",
     "seeMore": "See more",
     "whyDecline": "Why would you like to decline?",
     "whyDeclineSubtitle": "Your reason is private and will not be shared with {name}.",
